requirements/base.txt
Django==2.2.6
djangorestframework==3.10.3
orjson==2.1.2
